refactor: move OutputTokenValue component to PoolOperations
