+++ src/components/PoolOperations/Form/InputTokens.tsx
@@ -6,8 +6,8 @@ import { ITokenDetails } from '../../../context/PoolTokensContext'
 
 import useMatomoEcommerce from '../../../hooks/useMatomoEcommerce'
 
-import InputTokenValue from '../../InputTokenValue'
-import OutputTokenValue from '../../OutputTokenValue/OutputTokenValue'
+import InputTokenValue from '../InputTokenValue'
+import OutputTokenValue from '../OutputTokenValue'
 import SelectInput from '../../SelectInput'
 
 import { BNtoDecimal } from '../../../utils/numerals'
