+++ src/components/PoolOperations/OutputTokenValue/styles.ts
@@ -1,5 +1,5 @@
 import styled from 'styled-components'
-import theme from '../../styles/theme'
+import theme from '../../../styles/theme'
 
 interface IInputProps {
   bestValue?: boolean;
